Refactoring and cleanup for users
Move users internal DB into its own file
Move error definitions into its own file
Remove duplicate enable/disable operations
Fix actor context value key type, as per lint
